Split the project remover for better testability
