+++ src/main/java/com/lesfurets/maven/partial/mojos/ChangedMojo.java
@@ -82,9 +82,9 @@ public class ChangedMojo extends AbstractMojo {
         }
 
         Injector injector = Guice.createInjector(new GuiceModule(new MavenToPlexusLogAdapter(getLog()), session));
-        UnchangedProjectsRemover projectsRemover = injector.getInstance(UnchangedProjectsRemover.class);
         ChangedProjects changedProjects = injector.getInstance(ChangedProjects.class);
         Configuration configuration = injector.getInstance(Configuration.class);
+        ImpactedProjects impactedProjects = injector.getInstance(ImpactedProjects.class);
 
         getLog().info(configuration.toString());
 
@@ -92,11 +92,7 @@ public class ChangedMojo extends AbstractMojo {
             Set<MavenProject> changed = changedProjects.get();
             Set<MavenProject> ignoredProjects = configuration.ignoredProjects;
             changed.removeAll(ignoredProjects);
-            projectsRemover.collectDependentProjects(changed);
-
-            List<MavenProject> sortedChanged = session.getProjects().stream()
-                            .filter(changed::contains)
-                            .collect(Collectors.toList());
+            List<MavenProject> sortedChanged = impactedProjects.get(changed);
 
             PluginUtils.writeChangedProjectsToFile(sortedChanged, new File(outputFile));
             session.getProjects().forEach(m -> m.getProperties()
